--- hack/patch/0000_00_clusters.open-cluster-management.io_managedclustersets.crd.yaml.yaml-patch
@@ -1,16 +1,0 @@
-#This is yaml-patch config file. It's used to patch conversion webhook config to clusterset crd
-- op: add
-  path: /spec/conversion
-  value:
-    strategy: Webhook
-    webhook:
-      clientConfig:
-        service:
-          namespace: {{ .ClusterManagerNamespace }}
-          name: cluster-manager-registration-webhook
-          path: /convert
-          port: {{.RegistrationWebhook.Port}}
-        caBundle: {{ .RegistrationAPIServiceCABundle }}
-      conversionReviewVersions:
-      - v1beta1
-      - v1beta2
