--- pkg/registration/webhook/v1beta1/managedclusterset_conversion.go
@@ -1,4 +1,0 @@
-package v1beta1
-
-// Hub marks this type as a conversion hub.
-func (r *ManagedClusterSet) Hub() {}
