+++ pkg/registration/webhook/v1beta2/managedclustersetbinding_validating.go
@@ -2,15 +2,20 @@ package v1beta2
 
 import (
 	"context"
+	"fmt"
 
+	"k8s.io/api/apps/v1beta1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/client-go/kubernetes"
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
 	"open-cluster-management.io/api/cluster/v1beta2"
 
-	internalv1beta1 "open-cluster-management.io/ocm/pkg/registration/webhook/v1beta1"
+	authenticationv1 "k8s.io/api/authentication/v1"
+	authorizationv1 "k8s.io/api/authorization/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 var _ webhook.CustomValidator = &ManagedClusterSetBindingWebhook{}
@@ -32,7 +37,7 @@ func (b *ManagedClusterSetBindingWebhook) ValidateCreate(ctx context.Context, ob
 	if err != nil {
 		return nil, apierrors.NewBadRequest(err.Error())
 	}
-	return nil, internalv1beta1.AllowBindingToClusterSet(b.kubeClient, binding.Spec.ClusterSet, req.UserInfo)
+	return nil, AllowBindingToClusterSet(b.kubeClient, binding.Spec.ClusterSet, req.UserInfo)
 }
 
 // ValidateUpdate implements webhook.Validator so a webhook will be registered for the type
@@ -54,4 +59,44 @@ func (b *ManagedClusterSetBindingWebhook) ValidateUpdate(ctx context.Context, ol
 func (b *ManagedClusterSetBindingWebhook) ValidateDelete(_ context.Context, obj runtime.Object) (
 	admission.Warnings, error) {
 	return nil, nil
+}
+
+// allowBindingToClusterSet checks if the user has permission to bind a particular cluster set
+func AllowBindingToClusterSet(kubeClient kubernetes.Interface, clusterSetName string, userInfo authenticationv1.UserInfo) error {
+	extra := make(map[string]authorizationv1.ExtraValue)
+	for k, v := range userInfo.Extra {
+		extra[k] = authorizationv1.ExtraValue(v)
+	}
+
+	sar := &authorizationv1.SubjectAccessReview{
+		Spec: authorizationv1.SubjectAccessReviewSpec{
+			User:   userInfo.Username,
+			UID:    userInfo.UID,
+			Groups: userInfo.Groups,
+			Extra:  extra,
+			ResourceAttributes: &authorizationv1.ResourceAttributes{
+				Group:       "cluster.open-cluster-management.io",
+				Resource:    "managedclustersets",
+				Subresource: "bind",
+				Verb:        "create",
+				Name:        clusterSetName,
+			},
+		},
+	}
+	sar, err := kubeClient.AuthorizationV1().SubjectAccessReviews().Create(context.TODO(), sar, metav1.CreateOptions{})
+	if err != nil {
+		return apierrors.NewForbidden(
+			v1beta1.Resource("managedclustersets/bind"),
+			clusterSetName,
+			err,
+		)
+	}
+	if !sar.Status.Allowed {
+		return apierrors.NewForbidden(
+			v1beta1.Resource("managedclustersets/bind"),
+			clusterSetName,
+			fmt.Errorf("user %q is not allowed to bind cluster set %q", userInfo.Username, clusterSetName),
+		)
+	}
+	return nil
 }
